requirements.txt
pycryptodome==3.20.0
setuptools==69.5.1
wheel==0.43.0
